* commit 'd66410b0acd3889570b7ba7bd86facbc46e83b41':
  build: rework version reporting

+++ provider/plugins/DBUserPlugin.h
@@ -190,7 +190,7 @@ public:
 extern "C" {
 	extern _kc_export UserPlugin *getUserPluginInstance(std::mutex &, ECPluginSharedData *);
 	extern _kc_export void deleteUserPluginInstance(UserPlugin *);
-	extern _kc_export int getUserPluginVersion(void);
+	extern _kc_export unsigned long getUserPluginVersion(void);
 }
 /** @} */
